enhance(super-stream): add filter parameter to declareSuperStreamConsumer and SuperStreamConsumer for enhanced message filtering capabilities
